refactor: rename folder to dto
